// api/src/database/database.types.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export interface Tbformaspago {
  c_formapago: string;
  c_moneda: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  fp: number;
  id: Generated<number>;
  nomfp: string;
  usumod: number;
  usunvo: number;
}